Extract scripts to separate file
